first commit: attempt funky typemap option
In order to make this ergonomic I think the user should be able to
provide a type and YAAP should handle parsing/converting it and
panicking if necessary (I think clap makes you do your own conversions).
But to do that, we need to maintain a collection of generic argument
types that when queried by name return their corresponding type. I don't
think there's a great way to do this. I think the best option is to
maintain a typemap to store all arguments that takes a value of each
given type and returns a HashMap<&str,ThatType>. This is pretty clunky
because it would mean implementing typemap::Key for all values that
should be convertible, which would require me to maintain a fork of the
project. There's other awkward boilerplate stuff I don't particularly
like either. So we're tabling this for now