funny misunderstanding: how String::starts_with(&[char]) works
I thought `String::starts_with(P)` where P is the char array variant
would check for characters in that order. That is, I thought
`s.starts_with(&['h','i'][..])` was the same as `s.starts_with("hi")`.
However, the former version just checks if the first character of `s` is
any of the chars in the array.

This was kind of disguised for a while because the case I was actually
using was a superset of the case I expected (if I expect characters 'a'
and 'b' in that order, then the first character is probably either 'a'
or 'b').

The solution is to instead use
"String::chars().zip(&['a','b',]).all(|(x,&y)| x==y)" instead