collect args for help message later
there are a couple ways to do this, none that great.
The least bad way would to be to maintain a vector of args as they're
used. But the generic-ness makes things unpleasant: I think it would
need to be a vector of boxed ArgTraits whose MatchType was generic over
a new trait (and it would require tweaking things, and I'm not sure it
would work).
Another not-terrible option would be to not store args at all. Instead,
as each arg is seen, compute and store the help statement. That would
mean doing a computation every time that is not needed in the common
case, though, which seems undesirable (even though it's not particularly
expensive).
The current solution is to strip the args of their generic-ness and
store that. This seems like a decent compromise, but is still a little
funky. idk, might revisit