The issue was with `free` args getting confused
All elements of argv are `free` to start out. If any argtype matches
against one of them, it makes it non-free, i.e. claimed. Any arguments
which are "free" at the end haven't been matched by any args.

Maybe instead of an `argv` and a `free` vec of Strings and bools, there
should just be an argv: Vec<ArgType7> with string and bool members.
Probably, huh?

I should be more creative with names. "Arg" currently means an input
from the commandline (e.g. cp arg1 arg2) or an Arg struct (e.g.
`Arg::from("dst", "the path to copy files to")`. `Match` is used all
over the place in different variants, just not "match" because it's a
keyword. language, man