remove dependence on Default trait
There are still some cases when we need to generate an object without
knowing what it should be. This is maybe an indication of poor design
and should be revisited.
For now it's not meaningfully different to use T::default() vs unsafe
zeroes: if either of them make any impact on anything then there's a
bug. Now at least we draw attention to it with the `unsafe`s

Should probably try to get rid of these at some point