switch up how ArgS::matches works
it seems like we should be able to safely subvert the borrowck here.
like with

```
{
let m = self.matches(long, short);
if m != ArgMatch::NoMatch { return m; }
}
self.used = true;
return ArgMatch::NoMatch
```

I know that's a little funky, but the alternative is to use `self.used =
true` wherever it comes up *before* the borrow happens, which is
awkward. And it would require short_matches and long_matches to both
borrow mutably, which is impossible because they return an enum that
might borrow into ArgS, and they sometimes need to both be called. This
is one of those times when Rust isn't letting us do something that is
definitely safe and we need to remove good abstractions (i.e. functions)
and use an uglier style in order to circumvent this. But at least it's
localized to one function and we don't need to actually subvert anything
(no unsafe or RefCell/Rc nonsense)