Clean up: Add example, fix #[must_use], remove YaapDone
Some day (tomorrow?) I'd like to soften the requirement of YaapArg on
FromStr, as I don't think the valid utf-8 requirement is necessary.
Paths are a common command-line arg, but they can be invalid utf-8.
Currently if they don't parse that's a panic. This is suboptimal.

Not sure what the best way to do this is. A negative trait could be
cool, where YaapArg is implemented for all FromStr, and for all
From<[u8]> where FromStr is not. But this doesn't seem promising. Should
look into OIBIT workarounds/solution or maybe specialization? Or like
grep the source for "impl<T: From<Vec<u8>>> FromStr for T" but I think
that would be too easy.

So goes the grind.