add list impl and tests ("100%" coverage)
next look into subcmds and extra examples

also look into how to add a `.is_required()` method to more than one
ArgM<ArgTrait>. For now, it's only on ValArg, but adding it to ListArg
causes the compiler to not be able to infer types properly. This can
probably be worked around by adding a `is_required()` method to
ArgTrait, but then requiring a FlagArg wouldn't be a compile-time error
anymore. Maybe I can create a new trait Requirable that is impl'd for
everything *except* FlagArg?

if not I think I'm okay with ValArg being the only requirable option. It
should definitely be requirable because the distinction is type-level
(the difference is between returning a `T` and an `Option<T>`), but the
alternatives are value-level (an unused CountArg returns `0`, an unused
ListArg returns `vec![]`). These are easy enough to check at runtime and
have extremely sane empty values, which I'm okay with