Alternate ValOptArg doesn't work too well
Ideally a ValArg would only be created by a ValOptArg on which
.required() or .default(..) is called. But this means the following
should not be allowed: `.get_val(&mut T, Arg::new(..))` because it
creates a new ValArg via invalid means.

In this case, a plain old ValArg is paired with a `&mut T`. If the
ValArg isn't made with .required or .default then it is possible that it
is neither required nor has a default. In which case well-formed input
might cause `T` to never be set. Ideally we'd maintain the invariant
that all `result` inputs are set, so that it is safe for users to use
::std::mem::zeroed() to create arg objects. This violates that guarantee
by sometimes not setting `result: &mut T`

The ideal way to fix this is to only implement `new()` for ArgM variants
which can be created organically. But we can't implement `new()` 4
different times or we break the type inference. Maybe a `trait
Creatable` could do this? That would actually be kinda neat.